refactor: requirements components
